requirements.txt
requests>=2.28.0
pillow>=9.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pytest>=7.0.0
pytest-asyncio>=0.21.0